+++ gridcomps/History/Sampler/MAPL_EpochSwathMod.F90
@@ -58,6 +58,7 @@ module MAPL_EpochSwathMod
      procedure :: fill_time_in_bundle
      procedure :: find_config
      procedure :: config_accumulate
+     procedure :: verify_epoch_equals_freq
   end type samplerHQ
 
   interface samplerHQ
@@ -92,8 +93,8 @@ module MAPL_EpochSwathMod
      logical :: itemOrderAlphabetical = .true.
      integer :: fraction
      logical :: have_initalized
+     integer :: epoch_sec
    contains
-!!     procedure :: CreateFileMetaData
      procedure :: Create_bundle_RH
      procedure :: CreateVariable
      procedure :: regridScalar
@@ -103,7 +104,6 @@ module MAPL_EpochSwathMod
      procedure :: check_chunking
      procedure :: alphabatize_variables
      procedure :: addVariable_to_acc_bundle
-     procedure :: addVariable_to_output_bundle
      procedure :: interp_accumulate_fields
   end type sampler
 
@@ -191,6 +191,39 @@ contains
     this%CF_loc(this%ngrid)%cf = cf
     _RETURN(_SUCCESS)
   end subroutine config_accumulate
+
+
+  subroutine verify_epoch_equals_freq (this, frequency_from_list, swath_grid_label, rc)
+    class(samplerHQ)  :: this
+    integer, intent(in) :: frequency_from_list
+    character(len=*) , intent(in) :: swath_grid_label
+    integer, intent(out), optional :: rc
+    type(ESMF_Config) :: config_grid
+    integer :: hq_epoch_sec
+    integer :: freq_sec
+    integer :: local_swath_epoch_sec
+    integer :: time_integer
+    logical :: con
+    integer :: status
+
+    call ESMF_TimeIntervalGet(this%Frequency_epoch, s=hq_epoch_sec, _RC)
+    freq_sec = MAPL_nsecf( frequency_from_list )
+    config_grid = this%find_config( swath_grid_label )
+    call ESMF_ConfigGetAttribute(config_grid, value=time_integer, &
+         label=trim(swath_grid_label)//'.Epoch:', default=0, _RC)
+    local_swath_epoch_sec = MAPL_nsecf( time_integer )
+
+    con = (hq_epoch_sec == local_swath_epoch_sec) .AND. (hq_epoch_sec == freq_sec)
+    if (mapl_am_i_root()) then
+       if (.not. con) then
+          write(6, '(2x,a,2x,i10)') 'hq_epoch_sec', hq_epoch_sec
+          write(6, '(2x,a,2x,i10)') 'local_swath_epoch_sec', local_swath_epoch_sec
+          write(6, '(2x,a,2x,i10)') 'freq_sec', freq_sec
+       end if
+    end if
+    _ASSERT(con, 'Error in '//trim(swath_grid_label)//' related swath and list in History.rc: Epoch in all swath grids must be equal, and equal to list%freq')
+    _RETURN(_SUCCESS)
+  end subroutine verify_epoch_equals_freq
 
 
   !--------------------------------------------------!
@@ -312,22 +345,24 @@ contains
     !__ s3. destroy acc_bundle / output_bundle
 
    call ESMF_FieldBundleGet(sp%acc_bundle,fieldCount=numVars,_RC)
-   allocate(names(numVars),stat=status)
+   allocate(names(numVars),_STAT)
    call ESMF_FieldBundleGet(sp%acc_bundle,fieldNameList=names,_RC)
    do i=1,numVars
       call ESMF_FieldBundleGet(sp%acc_bundle,trim(names(i)),field=field,_RC)
       call ESMF_FieldDestroy(field,noGarbage=.true., _RC)
    enddo
    call ESMF_FieldBundleDestroy(sp%acc_bundle,noGarbage=.true.,_RC)
+   deallocate(names,_STAT)
 
    call ESMF_FieldBundleGet(sp%output_bundle,fieldCount=numVars,_RC)
-   allocate(names(numVars),stat=status)
+   allocate(names(numVars),_STAT)
    call ESMF_FieldBundleGet(sp%output_bundle,fieldNameList=names,_RC)
    do i=1,numVars
       call ESMF_FieldBundleGet(sp%output_bundle,trim(names(i)),field=field,_RC)
       call ESMF_FieldDestroy(field,noGarbage=.true., _RC)
    enddo
    call ESMF_FieldBundleDestroy(sp%output_bundle,noGarbage=.true.,_RC)
+   deallocate(names,_STAT)
 
    _RETURN(ESMF_SUCCESS)
 
@@ -439,7 +474,7 @@ contains
         if (allocated(this%metadata)) then
            deallocate (this%metadata)
         end if
-        allocate(this%metadata)
+        allocate(this%metadata,_STAT)
         call factory%append_metadata(this%metadata)
         if (present(vdata)) then
            this%vdata=vdata
@@ -454,23 +489,22 @@ contains
         if (this%vdata%regrid_type == VERTICAL_METHOD_ETA2LEV) call this%vdata%get_interpolating_variable(this%input_bundle,rc=status)
         _VERIFY(status)
 
+        ! __ add field to output_bundle
+        !
         iter = this%items%begin()
         do while (iter /= this%items%end())
            item => iter%get()
            if (item%itemType == ItemTypeScalar) then
-              call this%CreateVariable(item%xname,rc=status)
-              _VERIFY(status)
+              call this%CreateVariable(item%xname,_RC)
            else if (item%itemType == ItemTypeVector) then
-              call this%CreateVariable(item%xname,rc=status)
-              _VERIFY(status)
-              call this%CreateVariable(item%yname,rc=status)
-              _VERIFY(status)
+              call this%CreateVariable(item%xname,_RC)
+              call this%CreateVariable(item%yname,_RC)
            end if
            call iter%next()
         enddo
 
 
-        ! __ add acc_bundle and output_bundle
+        ! __ add field to acc_bundle
         !
         this%acc_bundle = ESMF_FieldBundleCreate(_RC)
         call ESMF_FieldBundleSet(this%acc_bundle,grid=this%output_grid,_RC)
@@ -519,8 +553,7 @@ contains
         if (present(quantize_algorithm)) this%quantizeAlgorithm = quantize_algorithm
         if (present(quantize_level)) this%quantizeLevel = quantize_level
         if (present(chunking)) then
-           allocate(this%chunking,source=chunking,stat=status)
-           _VERIFY(status)
+           allocate(this%chunking,source=chunking,_STAT)
         end if
         if (present(itemOrder)) this%itemOrderAlphabetical = itemOrder
         if (present(write_collection_id)) this%write_collection_id=write_collection_id
@@ -538,14 +571,14 @@ contains
         call MAPL_GridGet(this%output_grid,globalCellCountPerDim=global_dim,rc=status)
         _VERIFY(status)
         if (global_dim(1)*6 == global_dim(2)) then
-           allocate(this%chunking(5))
+           allocate(this%chunking(5),_STAT)
            this%chunking(1) = global_dim(1)
            this%chunking(2) = global_dim(1)
            this%chunking(3) = 1
            this%chunking(4) = 1
            this%chunking(5) = 1
         else
-           allocate(this%chunking(4))
+           allocate(this%chunking(4),_STAT)
            this%chunking(1) = global_dim(1)
            this%chunking(2) = global_dim(2)
            this%chunking(3) = 1
@@ -564,8 +597,7 @@ contains
         integer :: status
         character(len=5) :: c1,c2
 
-        call MAPL_GridGet(this%output_grid,globalCellCountPerDim=global_dim,rc=status)
-        _VERIFY(status)
+        call MAPL_GridGet(this%output_grid,globalCellCountPerDim=global_dim,_RC)
         if (global_dim(1)*6 == global_dim(2)) then
            write(c2,'(I5)')global_dim(1)
            write(c1,'(I5)')this%chunking(1)
@@ -693,10 +725,9 @@ contains
                  call ESMF_FieldGet(field,farrayPtr=ptr3d,rc=status)
                  _VERIFY(status)
               else
-                 allocate(ptr3d(0,0,0))
+                 allocate(ptr3d(0,0,0),_STAT)
               end if
-              allocate(ptr3d_inter(size(ptr3d,1),size(ptr3d,2),this%vdata%lm),stat=status)
-              _VERIFY(status)
+              allocate(ptr3d_inter(size(ptr3d,1),size(ptr3d,2),this%vdata%lm),_STAT)
               if (this%vdata%regrid_type==VERTICAL_METHOD_SELECT) then
                  call this%vdata%regrid_select_level(ptr3d,ptr3d_inter,rc=status)
                  _VERIFY(status)
@@ -725,13 +756,13 @@ contains
               call MAPL_FieldGetPointer(field,ptr2d,rc=status)
               _VERIFY(status)
            else
-              allocate(ptr2d(0,0))
+              allocate(ptr2d(0,0),_STAT)
            end if
            if (hasDE_out) then
               call MAPL_FieldGetPointer(OutField,outptr2d,rc=status)
               _VERIFY(status)
            else
-              allocate(outptr2d(0,0))
+              allocate(outptr2d(0,0),_STAT)
            end if
            if (gridIn==gridOut) then
               outPtr2d=ptr2d
@@ -752,14 +783,14 @@ contains
                  call ESMF_FieldGet(field,farrayPtr=ptr3d,rc=status)
                  _VERIFY(status)
               else
-                 allocate(ptr3d(0,0,0))
+                 allocate(ptr3d(0,0,0),_STAT)
               end if
            end if
            if (hasDE_out) then
               call MAPL_FieldGetPointer(OutField,outptr3d,rc=status)
               _VERIFY(status)
            else
-              allocate(outptr3d(0,0,0))
+              allocate(outptr3d(0,0,0),_STAT)
            end if
            if (gridIn==gridOut) then
               outPtr3d=Ptr3d
@@ -820,10 +851,9 @@ contains
                  call ESMF_FieldGet(xfield,farrayPtr=xptr3d,rc=status)
                  _VERIFY(status)
               else
-                 allocate(xptr3d(0,0,0))
+                 allocate(xptr3d(0,0,0),_STAT)
               end if
-              allocate(xptr3d_inter(size(xptr3d,1),size(xptr3d,2),this%vdata%lm),stat=status)
-              _VERIFY(status)
+              allocate(xptr3d_inter(size(xptr3d,1),size(xptr3d,2),this%vdata%lm),_STAT)
               if (this%vdata%regrid_type==VERTICAL_METHOD_SELECT) then
                  call this%vdata%regrid_select_level(xptr3d,xptr3d_inter,rc=status)
                  _VERIFY(status)
@@ -845,10 +875,9 @@ contains
                  call ESMF_FieldGet(yfield,farrayPtr=yptr3d,rc=status)
                  _VERIFY(status)
               else
-                 allocate(yptr3d(0,0,0))
+                 allocate(yptr3d(0,0,0),_STAT)
               end if
-              allocate(yptr3d_inter(size(yptr3d,1),size(yptr3d,2),this%vdata%lm),stat=status)
-              _VERIFY(status)
+              allocate(yptr3d_inter(size(yptr3d,1),size(yptr3d,2),this%vdata%lm),_STAT)
               if (this%vdata%regrid_type==VERTICAL_METHOD_SELECT) then
                  call this%vdata%regrid_select_level(yptr3d,yptr3d_inter,rc=status)
                  _VERIFY(status)
@@ -862,8 +891,7 @@ contains
               yptr3d => yptr3d_inter
            end if
         else
-           if (associated(xptr3d)) nullify(xptr3d)
-           if (associated(yptr3d)) nullify(yptr3d)
+           nullify(xptr3d, yptr3d)
         end if
 
         call ESMF_FieldBundleGet(this%input_bundle,xname,field=xfield,rc=status)
@@ -879,8 +907,8 @@ contains
               call MAPL_FieldGetPointer(yfield,yptr2d,rc=status)
               _VERIFY(status)
            else
-              allocate(xptr2d(0,0))
-              allocate(yptr2d(0,0))
+              allocate(xptr2d(0,0),_STAT)
+              allocate(yptr2d(0,0),_STAT)
            end if
 
            if (hasDE_in) then
@@ -889,8 +917,8 @@ contains
               call MAPL_FieldGetPointer(yOutField,youtptr2d,rc=status)
               _VERIFY(status)
            else
-              allocate(xoutptr2d(0,0))
-              allocate(youtptr2d(0,0))
+              allocate(xoutptr2d(0,0),_STAT)
+              allocate(youtptr2d(0,0),_STAT)
            end if
 
 
@@ -907,7 +935,7 @@ contains
                  call MAPL_FieldGetPointer(xfield,xptr3d,rc=status)
                  _VERIFY(status)
               else
-                 allocate(xptr3d(0,0,0))
+                 allocate(xptr3d(0,0,0),_STAT)
               end if
            end if
            if (.not.associated(yptr3d)) then
@@ -915,7 +943,7 @@ contains
                  call MAPL_FieldGetPointer(yfield,yptr3d,rc=status)
                  _VERIFY(status)
               else
-                 allocate(yptr3d(0,0,0))
+                 allocate(yptr3d(0,0,0),_STAT)
               end if
            end if
 
@@ -925,8 +953,8 @@ contains
               call MAPL_FieldGetPointer(yOutField,youtptr3d,rc=status)
               _VERIFY(status)
            else
-              allocate(xoutptr3d(0,0,0))
-              allocate(youtptr3d(0,0,0))
+              allocate(xoutptr3d(0,0,0),_STAT)
+              allocate(youtptr3d(0,0,0),_STAT)
            end if
 
            if (gridIn==gridOut) then
@@ -962,7 +990,7 @@ contains
      order = this%metadata%get_order(rc=status)
      _VERIFY(status)
      n = Order%size()
-     allocate(temp(nFixedVars+1:n))
+     allocate(temp(nFixedVars+1:n),_STAT)
      do i=1,n
         v1 => order%at(i)
         if ( i > nFixedVars) temp(i)=trim(v1)
@@ -991,7 +1019,7 @@ contains
      enddo
      call this%metadata%set_order(newOrder,rc=status)
      _VERIFY(status)
-     deallocate(temp)
+     deallocate(temp,_STAT)
 
      _RETURN(_SUCCESS)
 
@@ -1021,28 +1049,6 @@ contains
   end subroutine addVariable_to_acc_bundle
 
 
-  subroutine addVariable_to_output_bundle(this,itemName,rc)
-    class (sampler), intent(inout) :: this
-    character(len=*), intent(in) :: itemName
-    integer, optional, intent(out) :: rc
-
-    type(ESMF_Field) :: field,newField
-    integer :: fieldRank
-    integer :: status
-
-    call ESMF_FieldBundleGet(this%input_bundle,itemName,field=field,_RC)
-    call ESMF_FieldGet(field,rank=fieldRank,rc=status)
-    if (this%doVertRegrid .and. (fieldRank ==3) ) then
-       newField = MAPL_FieldCreate(field,this%output_grid,lm=this%vData%lm,_RC)
-    else
-       newField = MAPL_FieldCreate(field,this%output_grid,_RC)
-    end if
-    call MAPL_FieldBundleAdd(this%output_bundle,newField,_RC)
-
-    _RETURN(_SUCCESS)
-  end subroutine addVariable_to_output_bundle
-
-
 
   !! -- based on subroutine bundlepost(this,filename,oClients,rc)
   !!
@@ -1054,7 +1060,7 @@ contains
     integer, optional, intent(out) :: rc
 
     integer :: status
-    type(ESMF_Field) :: outField
+    type(ESMF_Field) :: outField, outField2
     type(ESMF_Field) :: new_outField
     type(ESMF_Grid)  :: grid
 
@@ -1092,11 +1098,11 @@ contains
 
     call ESMF_FieldBundleGet(this%output_bundle, grid=grid, _RC)
     call ESMF_GridGet(grid, localDECount=localDECount, dimCount=dimCount, _RC)
-    allocate ( LB(dimCount), UB(dimCount), exclusiveCount(dimCount) )
-    allocate ( compLB(dimCount), compUB(dimCount), compCount(dimCount) )
+    allocate ( LB(dimCount), UB(dimCount), exclusiveCount(dimCount) ,_STAT)
+    allocate ( compLB(dimCount), compUB(dimCount), compCount(dimCount) ,_STAT)
 
-    allocate ( j1(0:localDEcount-1) )  ! start
-    allocate ( j2(0:localDEcount-1) )  ! end
+    allocate ( j1(0:localDEcount-1) ,_STAT)  ! start
+    allocate ( j2(0:localDEcount-1) ,_STAT)  ! end
 
     _ASSERT ( localDEcount == 1, 'failed, due to localDEcount > 1')
     call MAPL_GridGetInterior(grid,ii1,iin,jj1,jjn)
@@ -1145,19 +1151,66 @@ contains
     do while (iter /= this%items%end())
        item => iter%get()
        if (item%itemType == ItemTypeScalar) then
-          call this%RegridScalar(item%xname,rc=status)
-          _VERIFY(status)
+          call this%RegridScalar(item%xname,_RC)
           call ESMF_FieldBundleGet(this%output_bundle,item%xname,field=outField, _RC)
-          _VERIFY(status)
           if (this%vdata%regrid_type==VERTICAL_METHOD_ETA2LEV) then
-             call this%vdata%correct_topo(outField,rc=status)
-             _VERIFY(status)
+             call this%vdata%correct_topo(outField,_RC)
           end if
+       elseif (item%itemType == ItemTypeVector) then
+          call this%RegridVector(item%xname,item%yname,_RC)
+          call ESMF_FieldBundleGet(this%output_bundle,item%xname,field=outField, _RC)
+          if (this%vdata%regrid_type==VERTICAL_METHOD_ETA2LEV) then
+             call this%vdata%correct_topo(outField,_RC)
+          end if
+          call ESMF_FieldBundleGet(this%output_bundle,item%yname,field=outField2, _RC)
+          if (this%vdata%regrid_type==VERTICAL_METHOD_ETA2LEV) then
+             call this%vdata%correct_topo(outField2,_RC)
+          end if
+       end if
 
-          ! -- mask the time interval
-          !    store the time interval fields into new bundle
-          call ESMF_FieldGet(outField, Array=array1, _RC)
-          call ESMF_FieldBundleGet(this%acc_bundle,item%xname,field=new_outField,_RC)
+
+       ! -- mask the time interval
+       !    store the time interval fields into new bundle
+       !    xname
+       call ESMF_FieldGet(outField, Array=array1, _RC)
+       call ESMF_FieldBundleGet(this%acc_bundle,item%xname,field=new_outField,_RC)
+       call ESMF_FieldGet(new_outField, Array=array2, _RC)
+       call ESMF_ArrayGet(array1, rank=rank, _RC)
+       if (rank==2) then
+          call ESMF_ArrayGet(array1, farrayptr=pt2d, _RC)
+          call ESMF_ArrayGet(array2, farrayptr=pt2d_, _RC)
+          localDe=0
+          if (j1(localDe)>0) then
+             do j= j1(localDe), j2(localDe)
+                jj= j-jj1+1     ! j_local
+                !!                      write(6,*) 'j, jj', j, jj
+                pt2d_(:,jj) = pt2d(:,jj)
+             enddo
+          endif
+       elseif (rank==3) then
+          call ESMF_ArrayGet(array1, farrayptr=pt3d, _RC)
+          call ESMF_ArrayGet(array2, farrayptr=pt3d_, _RC)
+          do localDe=0, localDEcount-1
+             if (j1(localDe)>0) then
+                do j= j1(localDe), j2(localDe)
+                   jj= j-jj1+1
+                   pt3d_(:,jj,:) = pt3d(:,jj,:)
+                enddo
+             endif
+          enddo
+       else
+          _FAIL('failed interp_accumulate_fields')
+       endif
+
+       ! __ additional step for yname if vector
+       if (item%itemType == ItemTypeScalar) then
+          ! already done
+       elseif (item%itemType == ItemTypeVector) then
+          !
+          ! add yname
+          !
+          call ESMF_FieldGet(outField2, Array=array1, _RC)
+          call ESMF_FieldBundleGet(this%acc_bundle,item%yname,field=new_outField,_RC)
           call ESMF_FieldGet(new_outField, Array=array2, _RC)
           call ESMF_ArrayGet(array1, rank=rank, _RC)
           if (rank==2) then
@@ -1185,9 +1238,6 @@ contains
           else
              _FAIL('failed interp_accumulate_fields')
           endif
-
-       else if (item%itemType == ItemTypeVector) then
-          _FAIL('ItemTypeVector not implemented')
        end if
        call iter%next()
     enddo